moved html-template-brick to one file
